+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 1.x branch
 ## 1.1 branch
 ### 1.1.9
+* `LinkScanner::_getResponse()` and `LinkScannerCommand::execute()` does not
+    suppress PHPUnit exceptions (e.g., the `Deprecated` exception), which are
+    throwned anyway;
 * ready for `cakephp` 4.3.
 
 ### 1.1.8